requirements.txt
# Trading Bot Dependencies
# Constitution v1.0.0 - §Dependencies: Pin all versions for reproducibility

# Core trading library
git+https://github.com/jmfernandes/robin_stocks.git@master  # Latest from GitHub with device approval auth fixes

# Data analysis
pandas==2.3.3
numpy==1.26.3
scipy==1.11.4  # Scientific computing (used for swing detection in TA framework)
yfinance==0.2.36  # Yahoo Finance data source for backtesting engine
pyarrow==22.0.0  # Parquet file format for efficient data caching (updated for PYSEC-2024-161)
# matplotlib==3.8.0  # P3 feature - visualization, deferred for future iteration

# Testing (§Testing_Requirements)
pytest==8.4.2
pytest-cov==4.1.0
pytest-mock==3.15.1
pytest-asyncio==1.2.0

# Type checking (§Code_Quality)
mypy==1.18.2
types-requests==2.28.11.5  # Downgraded for compatibility with polygon-api-client urllib3<2.0.0
types-pytz==2025.2.0.20250809  # Type stubs for pytz (safety-checks)

# Code quality (§Code_Quality)
ruff==0.14.1
pylint==3.0.3

# Security (§Security)
python-dotenv==1.0.0
bandit==1.8.6
pyotp==2.9.0  # MFA TOTP generation (authentication-module)

# Backtesting
backtrader==1.9.78.123
alpaca-py==0.43.2  # Alpaca Trading and Data API (paper trading for orchestrator)

# Utilities
python-dateutil==2.8.2
pytz==2024.1  # Timezone handling for trading hours (safety-checks)

# Dashboard UI (status-dashboard)
rich==13.7.0  # Terminal UI rendering
# pynput==1.8.1  # Keyboard input for dashboard controls (disabled - requires Linux headers, not needed in production)
PyYAML==6.0.3  # YAML config for dashboard targets

# CLI Tool (cli.py)
click==8.1.7  # Command-line interface framework
psutil==5.9.8  # Process monitoring and management
requests==2.31.0  # HTTP client for API calls

# Order Flow Monitoring (level-2-order-flow-i)
polygon-api-client==1.12.5  # Polygon.io SDK for Level 2 and Time & Sales data

# LLM-Friendly Bot Operations (029-llm-friendly-bot-operations)
fastapi==0.104.1  # REST API framework
uvicorn[standard]==0.24.0  # ASGI server with WebSocket support
pydantic==2.12.3  # Data validation (v2)
websockets==11.0.3  # Downgraded for compatibility with polygon-api-client
jsonschema==4.20.0  # JSON schema validation for config
sqlalchemy==2.0.23  # Database ORM for API order tracking
alembic==1.12.1  # Database migrations for API
openai==1.12.0  # OpenAI API client for LLM orchestrator
anthropic==0.39.0  # Anthropic Claude API client for multi-agent system
tenacity==8.2.3  # Retry library (dependency of openai)
tiktoken==0.5.2  # Token encoding for OpenAI models

# Telegram Notifications (030-telegram-notifications)
python-telegram-bot==20.7  # Async Telegram Bot API wrapper

# Telegram Command Handlers (031-telegram-command-handlers)
httpx~=0.25.2  # Async HTTP client for internal API calls (compatible with python-telegram-bot)

# Sentiment Analysis Integration (034-sentiment-analysis-integration)
transformers==4.35.0  # Hugging Face library for FinBERT model loading/inference
torch==2.6.0  # PyTorch for FinBERT model execution (upgraded to 2.6.0 for CVE-2025-32434 fix)
tweepy==4.14.0  # Twitter API v2 SDK for fetching tweets
praw==7.7.1  # Reddit API wrapper (Python Reddit API Wrapper)
